Update ts build to copy non-ts files; fix circular dep
