+++ src/common/routers/Utils.ts
@@ -21,6 +21,15 @@ import type { NextFunction, Request, Response, Router } from "express";
 import TagFormatter from "../TagFormatter";
 import { TaxNodeNotFoundError } from "../../routers/Search/adapters/common/ModelError";
 import Utils from "../Utils";
+import {
+  ContentTypeNotFoundError,
+  InvalidProjectIdError,
+  MissingParameterChooseError,
+  MissingParameterError,
+  ProjectNotFoundError,
+  TagNotFoundError,
+  UserError,
+} from "../UserError";
 
 // TODO: module or requirejs
 type ContentTypeRoute = (
@@ -337,73 +346,6 @@ export function handleUserErrors(logger: GmeLogger, ...fns: WebgmeHandler[]) {
       }
     }
   };
-}
-
-export class UserError extends Error {
-  statusCode: number;
-
-  constructor(msg: string, code = 400) {
-    super(msg);
-    this.statusCode = code;
-  }
-
-  sendVia(response: Response) {
-    response.status(this.statusCode).send(this.message);
-  }
-}
-
-class ProjectNotFoundError extends UserError {
-  constructor() {
-    super("Project not found", 404);
-  }
-}
-
-class ContentTypeNotFoundError extends UserError {
-  constructor(path: string) {
-    super(`Content type not found: ${path}`, 404);
-  }
-}
-
-class InvalidProjectIdError extends UserError {
-  constructor(projectId: string | undefined) {
-    const msg = projectId
-      ? `Invalid project ID: ${projectId}`
-      : `Project ID required.`;
-    super(msg);
-  }
-}
-
-class TagNotFoundError extends UserError {
-  constructor(tag: string) {
-    super(`Tag not found: ${tag}`, 404);
-  }
-}
-
-export class TestEnvOnlyError extends UserError {
-  constructor(name: string) {
-    super(`${name} is only allowed on a test deployment`, 403);
-  }
-
-  static check(name: string) {
-    if (typeof process !== "undefined" && process.env.ENV !== "test") {
-      throw new TestEnvOnlyError(name);
-    }
-  }
-}
-
-class MissingParameterError extends UserError {
-  constructor(param: string) {
-    super(`Missing URL parameter: ${param}`);
-  }
-}
-
-/**
- * Missing a required parameter. Must choose one from the given set.
- */
-class MissingParameterChooseError extends UserError {
-  constructor(...choices: string[]) {
-    super(`Must specify a URL parameter from: ${choices.join(" ")}`);
-  }
 }
 
 interface Emitter {
